+++ client/views/_views.scss
@@ -23,11 +23,6 @@
 @import 'find/find';
 @import 'find/filter/find.filter';
 
-// Frames
-@import 'frames/events/frame.events';
-@import 'frames/schedule/schedule';
-@import 'frames/week/frame.week';
-
 // Groups
 @import 'groups/details/group.details';
 @import 'groups/list/group.list';
