+++ imports/ui/layouts/frame.js
@@ -1,1 +1,3 @@
 import '/imports/ui/components/messages/messages.js';
+
+import './frame.html';
