+++ imports/ui/layouts/index.js
@@ -1,0 +1,1 @@
+import './frame.js';
