+++ imports/ui/pages/frames/propose/propose.js
@@ -1,0 +1,1 @@
+import './propose.html';
