+++ imports/ui/pages/index.scss
@@ -1,0 +1,7 @@
+@import './calendar/calendar';
+@import './faq/faq';
+@import './frames/events/events';
+@import './frames/schedule/schedule';
+@import './frames/week/week';
+@import './not-found/not-found';
+@import './timetable/timetable';
