+++ server/extracts.msgfmt~
@@ -1,1 +1,1 @@
-[{"translate.windowtitle":{"key":"translate.windowtitle","text":"Translate","file":"imports/ui/components/translate-info/translate-info.js","line":9,"mtime":1509466506727,"template":"unknown","ctime":1509466506727}},{"extractedAt":1509466514839,"updatedAt":1509466506727}]
+[{"course.list.windowtitle":{"key":"course.list.windowtitle","text":"Courses","file":"imports/ui/pages/frames/courselist/courselist.js","line":9,"mtime":1509468571883,"template":"unknown","ctime":1509468571883}},{"extractedAt":1509468580711,"updatedAt":1509468571883}]
